Split DataStore to seperate file
